Adding an info.plist for the shared spu library on OS X. Silence a few warnings about long to int conversions on OS X code.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@86929 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/plugins/DFInput/macsrc/ControllerList.m
@@ -125,7 +125,7 @@ static const int DPad[DKEY_TOTAL] = {
 			GetKeyDescription(buf, currentController, DPad[rowIndex]);
 		} else {
 			rowIndex -= DKEY_TOTAL;
-			GetAnalogDescription(buf, currentController, rowIndex / 4, rowIndex % 4);
+			GetAnalogDescription(buf, currentController, (int)(rowIndex / 4), rowIndex % 4);
 		}
 
 		return @(buf);
